Add Intro component to About page and refactor layout
